+++ R/sandbox.R
@@ -1,10 +1,37 @@
 
 renv_sandbox_activate <- function(project = NULL) {
 
+  # record start time
+  before <- Sys.time()
+
   # attempt the activation
   status <- catch(renv_sandbox_activate_impl(project))
   if (inherits(status, "error"))
     warning(status)
+
+  # record end time
+  after <- Sys.time()
+
+  # check for long elapsed time
+  elapsed <- difftime(after, before, units = "secs")
+
+  # if it took more than 30 seconds to activate the sandbox, warn the user
+  if (elapsed > 30) {
+
+    fmt <- paste(
+      "renv took longer than expected (%s) to activate the sandbox.",
+      "",
+      "The sandbox can be disabled by setting:",
+      "",
+      "    RENV_CONFIG_SANDBOX_ENABLED = FALSE",
+      "",
+      "within an appropriate start-up .Renviron file. See `?renv::config` for more details.",
+      sep = "\n"
+    )
+
+    warningf(fmt, renv_difftime_format(elapsed))
+
+  }
 
   # return status
   status